+++ ejemplos/proyectos/planetas-pymcel.py
@@ -1,0 +1,38 @@
+import pymcel as pc
+import numpy as np
+np.random.seed(42)
+
+M_central = 1.0
+N_planetesimales = 50
+masa_total_disco = 0.1 
+
+print(f"Generando disco protoplanetario con {N_planetesimales} planetesimales...")
+radio_minimo = 0.5
+radio_maximo = 3.0
+masas, pos, vel, radios = pc.condiciones_iniciales_planetesimales(
+    M_central=M_central,
+    N_planetesimales=N_planetesimales,
+    masa_total_disco=masa_total_disco,
+    radio_minimo=radio_minimo,
+    radio_maximo=radio_maximo
+)
+
+radios[0]*=0.1
+pc.ncuerpos_rebound_tiempo_real(
+    masas=masas, 
+    posiciones=pos, 
+    velocidades=vel, 
+    t_final=None, 
+    dt_grafico=0.1,
+    limite_grafico=5*radio_maximo,
+    titulo="Formación Planetaria (Acreción)",
+    plot_3d=False, 
+    recentrado=1,
+    trazos=0,
+    radios=radios,
+    colisiones=True,
+    tamanos_dinamicos=True,
+    i_central=0,
+    alpha_radio=1,
+    grabar_posiciones=True
+)